Fixes eff:div/0 bug on shepherd failures
If the shepherd fails while the job is starting, the slots column is set to zero.

Without this fix, that means the efficiency calculation has a 0 in the denominator, which causes a div/zero error.

The fix sets the slots to a minimum of 1. The numerator should be zero in this case, so the value will be zero in any case, but it won't cause an error.